refactor: consolidate split test files into one file per source class
Enforces the convention of one test file per production class. Eight
source files are deleted and their tests folded into the canonical test
for the class under test.

Consolidations
--------------
* AvroConverterTypeMappingTest → AvroConverterTest
  Parameterized primitive type-mapping tests (Rel→Avro and Avro→Rel)
  merged in. Adds ParameterizedTest / MethodSource / RelProtoDataType
  imports to AvroConverterTest.

* K8sViewTableRowTest → K8sViewTableTest
  Inner-class Row method tests (viewPath, schemaPath, viewName,
  toString) merged in. No new imports required.

* SubscriptionReconcilerFetchAttributesTest → SubscriptionReconcilerTest
  All reflection removed. The three tests that verified fetchAttributes()
  swallows malformed YAML are rewritten to call reconcile() with the
  problematic YAML placed in status.jobResources, exercising the same
  error-handling paths through the public API. Adds a
  buildPhase3Subscription() helper to avoid repetition.

* ScriptImplementorInnerClassesTest → ScriptImplementorTest (util)
  39 inner-class tests (Catalog, Database, Connector, CompoundIdentifier,
  Identifier, Seal, View, Statement, Query, Insert, DropFields, nullable
  column DDL) merged in. testInsertWithSuffix renamed to
  testInsertWithSuffixNullTargetFields to avoid collision with the
  existing test that covers the non-null targetFields path.

* catalog/ScriptImplementorTest → catalog/HopTableTest (rename)
  The catalog ScriptImplementorTest was testing HopTable.implement(),
  not ScriptImplementor (which lives in hoptimator-util). Renamed to
  HopTableTest to reflect the class actually under test.

Schema validation cleanup
-------------------------
KafkaSchemaValidationTest, MySQLSchemaValidationTest, and
VeniceSchemaValidationTest are deleted. Each is fully superseded by a
quidem .id test that uses the !describe command added in #203:
  kafka-ddl.id, mysql-ddl.id, venice-ddl-insert-partial.id
The quidem tests cover the same schema assertions and additionally
verify INSERT operations and deployment specs.

Net: 8 files deleted, 5 modified/created; -366 lines, no coverage loss.
