+++ hoptimator-catalog/src/test/java/com/linkedin/hoptimator/catalog/HopTableTest.java
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
-public class ScriptImplementorTest {
+public class HopTableTest {
 
   @Test
   public void implementsFlinkCreateTableDDL() {
