--- hoptimator-venice/src/test/java/com/linkedin/hoptimator/venice/VeniceSchemaValidationTest.java
@@ -1,33 +1,0 @@
-package com.linkedin.hoptimator.venice;
-
-import com.linkedin.hoptimator.jdbc.JdbcTestBase;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-import java.util.Map;
-
-
-/**
- * Tests Venice table creation and schema validation without requiring all drivers
- * to be loaded via metadata system queries.
- */
-@Tag("integration")
-public class VeniceSchemaValidationTest extends JdbcTestBase {
-
-    @Test
-    public void testVeniceTableCreationAndSchema() throws Exception {
-        sql("create or replace materialized view \"VENICE\".\"test-store$insert-partial\" (\"KEY_name\", \"KEY_age\", \"intField\") "
-            + "as select \"stringField\" AS \"KEY_name\", \"KEY\" AS \"KEY_age\", \"intField\" from \"VENICE\".\"test-store-primitive\"");
-
-        // Validate the table was created with expected schema
-        Map<String, String> expectedColumns = Map.of(
-            "KEY_name", "VARCHAR",
-            "KEY_age", "INTEGER",
-            "intField", "INTEGER"
-        );
-        validateTableSchema(List.of("VENICE", "test-store$insert-partial"), expectedColumns);
-
-        sql("drop materialized view venice.\"test-store$insert-partial\"");
-    }
-}
